+++ Makie/test/runtests.jl
@@ -58,6 +58,7 @@ end
         include("SceneLike/figures.jl")
         include("SceneLike/makielayout.jl")
         include("SceneLike/PolarAxis.jl")
+        include("events_isolation.jl")
     end
 
     @testset "Conversion & Projection Pipeline" begin